Refactor node events into [action]Node and node[action]ed.
This allows handlers that do the work from those that need to be notified when the work is done. The workspace would add nodes and graph was updating nodes before the app had added them to the workspace, so changes were not being persisted.

Also includes fixes for retina screens